refactor: move infraUtils.Retry into utils.Retry (#693)
