+++ infra/aliyun/ali_vpc.go
@@ -17,8 +17,7 @@ package aliyun
 import (
 	"errors"
 
-	"github.com/alibaba/sealer/infra/utils"
-
+	"github.com/alibaba/sealer/utils"
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests"
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/responses"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
